chore: split the HR data store to multiple files
