+++ edge-apps/bamboo-hr-app/src/types/employee.ts
@@ -1,0 +1,7 @@
+export interface Employee {
+  id: number
+  firstName: string
+  lastName: string
+  dateOfBirth: string
+  hireDate: string
+}
